requirements.txt
# === Dependencias del Bot de Telegram y Servidor API ===

# El framework principal para el bot de Telegram
python-telegram-bot

# Para el servidor web que recibe los dibujos
fastapi
uvicorn[standard]

# Para cargar las variables de entorno desde el archivo .env
python-dotenv


# === Dependencias para la IA en la Nube (OpenRouter) ===

# Librería oficial de OpenAI, compatible con la API de OpenRouter
openai


# === Dependencias para la IA Local (Hugging Face Transformers) ===

# La librería principal para ejecutar modelos de IA localmente
transformers

# El framework de Deep Learning sobre el que corre Transformers
# Nota: Esto instalará la versión para CPU. Si tienes una GPU NVIDIA con CUDA,
# deberías instalar la versión para GPU siguiendo las instrucciones de la web de PyTorch.
torch
torchvision
torchaudio

# Librería para el manejo y procesamiento de imágenes
Pillow

# Librería para descargar las imágenes de los objetivos desde sus URLs
requests